Add experimental Windows ARM64 CQ builder.
Along the way, it sets the vs_version and run_*_tests in the
configuration here. When this lands, we can remove the corresponding
logic from the recipe. It also reduces some duplicate configs for cases
when the CQ has a compile-only version of a CI config. (This, as a
result, ends up renaming some of the CQ builders to match their CI
counterparts.)

The CQ builder is, for now, disabled by default and not on CI. If it
works, I'll enable it.

(Once I've cleaned things up to use properties, I'll write some
documentation on all this.)

Bug: 397
Change-Id: I830b233bcbbbd145bacee9a2c9354305a9f3a923
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/45644
Reviewed-by: Adam Langley <agl@google.com>