Generate full analogs to all the compile-only builders.
At the time we first set up our CQ (see link in attached bug), there
were no Android builders in the pool. This doesn't seem to be the case
now. As a first step to enabling those, autoatically generate full
analogs to all compile-only CQ builders. This makes them available in
the UI but not part of Commit-Queue +1 or +2. We'll be able to see if
the Android configs work.

Bug: 202
Change-Id: Id19bf92d9de6bd9b0c327bc6f0e984fb24533b77
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/53725
Reviewed-by: Adam Langley <agl@google.com>